+++ Cargo.toml
@@ -359,6 +359,9 @@ tracing-core = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 trie-db = "0.27.1"
 typenum = { version = "1.14.0", default-features = false }
+tonic = { version = "0.9.2" }
+tonic-build = { version = "0.9.2", features = ["prost"] }
+tower = { version = "0.4" }
 untrusted = { version = "0.9.0" }
 vergen = { version = "8.2.6", default-features = false }
 wait-timeout = "0.2"
@@ -374,6 +377,7 @@ ces-node-rpc-ext-types = { path = "crates/ces-node-rpc-ext/types", default-featu
 ces-node-runtime = { path = "standalone/chain/runtime", package = "cess-node-runtime", default-features = false }
 ces-pallet-mq ={ path = "pallets/mq", default-features = false }
 ces-pallet-mq-runtime-api ={ path = "pallets/mq/runtime-api", default-features = false }
+ces-sanitized-logger = { path = "crates/ces-sanitized-logger" }
 ces-serde-more = { path = "crates/ces-serde-more", default-features = false }
 ces-trie-storage = { path = "crates/ces-trie-storage", default-features = false }
 ces-types = { path = "crates/ces-types", default-features = false }
@@ -386,8 +390,6 @@ cp-bloom-filter = { path = "crates/bloom-filter", default-features = false }
 cp-cess-common = { path = "crates/common", default-features = false }
 cp-enclave-verify = { path = "crates/enclave-verify", default-features = false }
 cp-scheduler-credit = { path = "crates/scheduler-credit", default-features = false }
-crpc = { path = "crates/crpc", default-features = false }
-crpc-build = { path = "crates/crpc-build" }
 ext-types = { path = "crates/ces-node-rpc-ext/types", default-features = false, package = "ces-node-rpc-ext-types" }
 node-executor = { path = "standalone/chain/executor" }
 node-rpc = { path = "standalone/chain/rpc" }
