+++ crates/cestory/api/build.rs
@@ -1,24 +1,15 @@
 fn main() {
-    use tera::{Context, Tera};
+    let mut builder = tonic_build::configure()
+        .out_dir("./src/proto_generated")
+        .disable_package_emission()
+        .build_server(true)
+        .build_client(false);
 
-    let tera = Tera::new("proto/*.proto").unwrap();
-
-    let tmpdir = tempdir::TempDir::new("rendered_proto").unwrap();
-    let render_dir = tmpdir.path();
-
-    for tmpl in tera.templates.keys() {
-        println!("cargo:rerun-if-changed=proto/{tmpl}");
-        let render_output = std::fs::File::create(render_dir.join(tmpl)).unwrap();
-        tera.render_to(tmpl, &Context::new(), render_output)
-            .unwrap();
+    #[cfg(feature = "ceseal-client")]
+    {
+        builder = builder.build_client(true);
     }
 
-    let out_dir = "./src/proto_generated";
-
-    let mut builder = crpc_build::configure()
-        .out_dir(out_dir)
-        .mod_prefix("crate::crpc::")
-        .disable_package_emission();
     builder = builder.type_attribute(
         ".ceseal_rpc",
         "#[derive(::serde::Serialize, ::serde::Deserialize)]",
@@ -31,7 +22,7 @@ fn main() {
     ] {
         builder = builder.type_attribute(name, "#[derive(::scale_info::TypeInfo)]");
     }
-    builder = builder.field_attribute("InitRuntimeResponse.attestation", "#[serde(skip,default)]");
+    builder = builder.field_attribute("InitRuntimeResponse.attestation", "#[serde(skip, default)]");
     for field in [
         "HttpFetch.body",
         "HttpFetch.headers",
@@ -39,6 +30,6 @@ fn main() {
         builder = builder.field_attribute(field, "#[serde(default)]");
     }
     builder
-        .compile(&["ceseal_rpc.proto"], &[render_dir])
+        .compile(&["ceseal_rpc.proto"], &["proto"])
         .unwrap();
 }
