+++ crates/cestory/api/src/lib.rs
@@ -1,13 +1,16 @@
 extern crate alloc;
 
-pub mod actions;
 pub mod blocks;
 pub mod crypto;
 pub mod ecall_args;
 pub mod endpoints;
 pub mod crpc;
-#[cfg(feature = "ceseal-client")]
-pub mod ceseal_client;
 pub mod storage_sync;
 
 mod proto_generated;
+
+#[cfg(feature = "ceseal-client")]
+pub mod ceseal_client {
+    use crate::crpc::ceseal_api_client::CesealApiClient;
+    pub type CesealClient<T> = CesealApiClient<T>;
+}
