--- crates/crpc-build/.rustfmt.toml
@@ -1,1 +1,0 @@
-hard_tabs = false
