--- crates/crpc/.rustfmt.toml
@@ -1,1 +1,0 @@
-hard_tabs = false
