--- crates/crpc/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "crpc"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-async-trait = { workspace = true }
-derive_more = { workspace = true }
-prost = { workspace = true, features = ["prost-derive"] }
-anyhow = { workspace = true }
-parity-scale-codec = { workspace = true }
-serde_json = { workspace = true, features = ["alloc"]}
-
-[features]
-default = ["std"]
-std = []
