+++ standalone/teeworker/cifrost/src/main.rs
@@ -1,4 +1,5 @@
 #[tokio::main]
 async fn main() {
+    ces_sanitized_logger::init_subscriber(false);
     cifrost::run().await;
 }
